Move collections to own packages
- Move local-only and local-storage collections to main @tanstack/db package
- Create new @tanstack/db-electric-collection package for Electric SQL integration
- Create new @tanstack/db-query-collection package for TanStack Query integration
- Delete @tanstack/db-collections package (to be deprecated on npm separately)
- Update example app and documentation to use new package structure

This reorganization provides better separation of concerns and allows each collection type to have independent versioning and dependencies.
