--- packages/db-collections/README.md
@@ -1,3 +1,0 @@
-# @tanstack/db-collections
-
-A collection of TanStack DB Collections. See [TanStack/db](https://github.com/TanStack/db) for more details.
