--- packages/db-collections/src/index.ts
@@ -1,23 +1,0 @@
-export {
-  electricCollectionOptions,
-  type ElectricCollectionConfig,
-  type ElectricCollectionUtils,
-  type Txid,
-} from "./electric"
-export {
-  queryCollectionOptions,
-  type QueryCollectionConfig,
-  type QueryCollectionUtils,
-} from "./query"
-export {
-  localStorageCollectionOptions,
-  type LocalStorageCollectionConfig,
-  type LocalStorageCollectionUtils,
-  type StorageApi,
-  type StorageEventApi,
-} from "./local-storage"
-export {
-  localOnlyCollectionOptions,
-  type LocalOnlyCollectionConfig,
-  type LocalOnlyCollectionUtils,
-} from "./local-only"
