--- packages/db-collections/tests/test-setup.ts
@@ -1,1 +1,0 @@
-// import "@testing-library/jest-dom/vitest"
