+++ packages/db-electric-collection/package.json
@@ -1,11 +1,10 @@
 {
-  "name": "@tanstack/db-collections",
-  "description": "A collection for (aspirationally) every way of loading your data",
-  "version": "0.0.23",
+  "name": "@tanstack/db-electric-collection",
+  "description": "Electric SQL collection for TanStack DB",
+  "version": "0.0.1",
   "dependencies": {
     "@standard-schema/spec": "^1.0.0",
     "@tanstack/db": "workspace:*",
-    "@tanstack/query-core": "^5.75.7",
     "@tanstack/store": "^0.7.0",
     "debug": "^4.4.1"
   },
@@ -43,10 +42,12 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/TanStack/db.git",
-    "directory": "packages/db-collections"
+    "directory": "packages/db-electric-collection"
   },
   "homepage": "https://tanstack.com/db",
   "keywords": [
+    "electric",
+    "sql",
     "optimistic",
     "typescript"
   ],
