+++ packages/db-electric-collection/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  electricCollectionOptions,
+  type ElectricCollectionConfig,
+  type ElectricCollectionUtils,
+  type Txid,
+} from "./electric"
