+++ packages/db-query-collection/src/index.ts
@@ -1,0 +1,5 @@
+export {
+  queryCollectionOptions,
+  type QueryCollectionConfig,
+  type QueryCollectionUtils,
+} from "./query"
