+++ packages/db-query-collection/vite.config.ts
@@ -1,0 +1,22 @@
+import { defineConfig, mergeConfig } from "vitest/config"
+import { tanstackViteConfig } from "@tanstack/config/vite"
+import packageJson from "./package.json"
+
+const config = defineConfig({
+  test: {
+    name: packageJson.name,
+    dir: `./tests`,
+    environment: `jsdom`,
+    coverage: { enabled: true, provider: `istanbul`, include: [`src/**/*`] },
+    typecheck: { enabled: true },
+    setupFiles: [`./tests/test-setup.ts`],
+  },
+})
+
+export default mergeConfig(
+  config,
+  tanstackViteConfig({
+    entry: `./src/index.ts`,
+    srcDir: `./src`,
+  })
+)
