+++ packages/db/src/index.ts
@@ -7,6 +7,8 @@ export * from "./errors"
 export * from "./proxy"
 export * from "./query/index.js"
 export * from "./optimistic-action"
+export * from "./local-only"
+export * from "./local-storage"
 
 // Re-export some stuff explicitly to ensure the type & value is exported
 export type { Collection } from "./collection"
